requirements.txt
django ~= 3.1.0
django-admin-interface ~= 0.13.5
django-cachalot ~= 2.3.1
django-ckeditor ~= 5.9.0
django-cors-headers ~= 3.4.0
django-crontab ~= 0.7.1
django-csp ~= 3.6
django-debug-toolbar ~= 2.2
django-json-widget ~= 1.0.1
django-environ ~= 0.4.5
django-htmlmin ~= 0.11.0
django-modeltranslation ~= 0.15.1
django-solo ~= 1.1.3
gevent ~= 20.6.2
gunicorn ~= 20.0.4
hashids ~= 1.3.1
mysqlclient ~= 2.0.1
python-benedict ~= 0.18.1
python-slugify ~= 4.0.1
requests ~= 2.24.0
sentry-sdk ~= 0.16.2
six ~= 1.14.0
whitenoise ~= 5.2.0